Combine NegoState into NegotiateStream
