+++ src/libraries/System.Net.Security/src/System/Net/Security/NegotiateStream.cs
@@ -2,13 +2,17 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System.ComponentModel;
+using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
 using System.IO;
-using System.Threading;
-using System.Threading.Tasks;
+using System.Runtime.CompilerServices;
+using System.Runtime.ExceptionServices;
+using System.Security.Authentication;
 using System.Security.Authentication.ExtendedProtection;
 using System.Security.Principal;
-using System.Diagnostics;
-using System.Runtime.CompilerServices;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace System.Net.Security
 {
@@ -17,9 +21,18 @@ namespace System.Net.Security
     /// </summary>
     public partial class NegotiateStream : AuthenticatedStream
     {
-        private readonly NegoState _negoState;
-        private readonly byte[] _readHeader;
+        private const int ERROR_TRUST_FAILURE = 1790;   // Used to serialize protectionLevel or impersonationLevel mismatch error to the remote side.
+        private const int MaxReadFrameSize = 64 * 1024;
+        private const int MaxWriteDataSize = 63 * 1024; // 1k for the framing and trailer that is always less as per SSPI.
+        private const string DefaultPackage = NegotiationInfoClass.Negotiate;
 
+#pragma warning disable CA1825 // used in reference comparison, requires unique object identity
+        private static readonly byte[] s_emptyMessage = new byte[0];
+#pragma warning restore CA1825
+        private static readonly AsyncCallback s_readCallback = new AsyncCallback(ReadCallback);
+        private static readonly AsyncCallback s_writeCallback = new AsyncCallback(WriteCallback);
+
+        private readonly byte[] _readHeader;
         private IIdentity? _remoteIdentity;
         private byte[] _buffer;
         private int _bufferOffset;
@@ -27,6 +40,23 @@ namespace System.Net.Security
 
         private volatile int _writeInProgress;
         private volatile int _readInProgress;
+        private volatile int _authInProgress;
+
+        private Exception? _exception;
+        private StreamFramer? _framer;
+        private NTAuthentication? _context;
+        private bool _canRetryAuthentication;
+        private ProtectionLevel _expectedProtectionLevel;
+        private TokenImpersonationLevel _expectedImpersonationLevel;
+        private uint _writeSequenceNumber;
+        private uint _readSequenceNumber;
+        private ExtendedProtectionPolicy? _extendedProtectionPolicy;
+
+        /// <summary>
+        /// SSPI does not send a server ack on successful auth.
+        /// This is a state variable used to gracefully handle auth confirmation.
+        /// </summary>
+        private bool _remoteOk = false;
 
         public NegotiateStream(Stream innerStream) : this(innerStream, false)
         {
@@ -34,7 +64,6 @@ namespace System.Net.Security
 
         public NegotiateStream(Stream innerStream, bool leaveInnerStreamOpen) : base(innerStream, leaveInnerStreamOpen)
         {
-            _negoState = new NegoState(innerStream);
             _readHeader = new byte[4];
             _buffer = Array.Empty<byte>();
         }
@@ -43,7 +72,8 @@ namespace System.Net.Security
         {
             try
             {
-                _negoState.Close();
+                _exception = new ObjectDisposedException(nameof(NegotiateStream));
+                _context?.CloseContext();
             }
             finally
             {
@@ -55,7 +85,8 @@ namespace System.Net.Security
         {
             try
             {
-                _negoState.Close();
+                _exception = new ObjectDisposedException(nameof(NegotiateStream));
+                _context?.CloseContext();
             }
             finally
             {
@@ -98,16 +129,16 @@ namespace System.Net.Security
             AsyncCallback? asyncCallback,
             object? asyncState)
         {
-            _negoState.ValidateCreateContext(NegoState.DefaultPackage, false, credential, targetName, binding, requiredProtectionLevel, allowedImpersonationLevel);
+            ValidateCreateContext(DefaultPackage, false, credential, targetName, binding, requiredProtectionLevel, allowedImpersonationLevel);
 
-            LazyAsyncResult result = new LazyAsyncResult(_negoState, asyncState, asyncCallback);
-            _negoState.ProcessAuthentication(result);
+            var result = new LazyAsyncResult(this, asyncState, asyncCallback);
+            ProcessAuthentication(result);
 
             return result;
         }
 
         public virtual void EndAuthenticateAsClient(IAsyncResult asyncResult) =>
-            _negoState.EndProcessAuthentication(asyncResult);
+            EndProcessAuthentication(asyncResult);
 
         public virtual void AuthenticateAsServer() =>
             AuthenticateAsServer((NetworkCredential)CredentialCache.DefaultCredentials, null, ProtectionLevel.EncryptAndSign, TokenImpersonationLevel.Identification);
@@ -120,8 +151,8 @@ namespace System.Net.Security
 
         public virtual void AuthenticateAsServer(NetworkCredential credential, ExtendedProtectionPolicy? policy, ProtectionLevel requiredProtectionLevel, TokenImpersonationLevel requiredImpersonationLevel)
         {
-            _negoState.ValidateCreateContext(NegoState.DefaultPackage, credential, string.Empty, policy, requiredProtectionLevel, requiredImpersonationLevel);
-            _negoState.ProcessAuthentication(null);
+            ValidateCreateContext(DefaultPackage, credential, string.Empty, policy, requiredProtectionLevel, requiredImpersonationLevel);
+            ProcessAuthentication(null);
         }
 
         public virtual IAsyncResult BeginAuthenticateAsServer(AsyncCallback? asyncCallback, object? asyncState) =>
@@ -146,16 +177,16 @@ namespace System.Net.Security
             AsyncCallback? asyncCallback,
             object? asyncState)
         {
-            _negoState.ValidateCreateContext(NegoState.DefaultPackage, credential, string.Empty, policy, requiredProtectionLevel, requiredImpersonationLevel);
+            ValidateCreateContext(DefaultPackage, credential, string.Empty, policy, requiredProtectionLevel, requiredImpersonationLevel);
 
-            LazyAsyncResult result = new LazyAsyncResult(_negoState, asyncState, asyncCallback);
-            _negoState.ProcessAuthentication(result);
+            var result = new LazyAsyncResult(this, asyncState, asyncCallback);
+            ProcessAuthentication(result);
 
             return result;
         }
 
         public virtual void EndAuthenticateAsServer(IAsyncResult asyncResult) =>
-            _negoState.EndProcessAuthentication(asyncResult);
+            EndProcessAuthentication(asyncResult);
 
         public virtual void AuthenticateAsClient() =>
             AuthenticateAsClient((NetworkCredential)CredentialCache.DefaultCredentials, null, string.Empty, ProtectionLevel.EncryptAndSign, TokenImpersonationLevel.Identification);
@@ -173,8 +204,8 @@ namespace System.Net.Security
         public virtual void AuthenticateAsClient(
             NetworkCredential credential, ChannelBinding? binding, string targetName, ProtectionLevel requiredProtectionLevel, TokenImpersonationLevel allowedImpersonationLevel)
         {
-            _negoState.ValidateCreateContext(NegoState.DefaultPackage, false, credential, targetName, binding, requiredProtectionLevel, allowedImpersonationLevel);
-            _negoState.ProcessAuthentication(null);
+            ValidateCreateContext(DefaultPackage, false, credential, targetName, binding, requiredProtectionLevel, allowedImpersonationLevel);
+            ProcessAuthentication(null);
         }
 
         public virtual Task AuthenticateAsClientAsync() =>
@@ -213,19 +244,52 @@ namespace System.Net.Security
             TokenImpersonationLevel requiredImpersonationLevel) =>
             Task.Factory.FromAsync((callback, state) => BeginAuthenticateAsServer(credential, policy, requiredProtectionLevel, requiredImpersonationLevel, callback, state), EndAuthenticateAsClient, null);
 
-        public override bool IsAuthenticated => _negoState.IsAuthenticated;
+        public override bool IsAuthenticated => IsAuthenticatedCore;
 
-        public override bool IsMutuallyAuthenticated => _negoState.IsMutuallyAuthenticated;
+        private bool IsAuthenticatedCore => _context != null && HandshakeComplete && _exception == null && _remoteOk;
 
-        public override bool IsEncrypted => _negoState.IsEncrypted;
+        public override bool IsMutuallyAuthenticated =>
+            IsAuthenticatedCore &&
+            !_context!.IsNTLM && // suppressing for NTLM since SSPI does not return correct value in the context flags.
+            _context.IsMutualAuthFlag;
 
-        public override bool IsSigned => _negoState.IsSigned;
+        public override bool IsEncrypted => IsAuthenticatedCore && _context!.IsConfidentialityFlag;
 
-        public override bool IsServer => _negoState.IsServer;
+        public override bool IsSigned => IsAuthenticatedCore && (_context!.IsIntegrityFlag || _context.IsConfidentialityFlag);
 
-        public virtual TokenImpersonationLevel ImpersonationLevel => _negoState.AllowedImpersonation;
+        public override bool IsServer => _context != null && _context.IsServer;
 
-        public virtual IIdentity RemoteIdentity => _remoteIdentity ??= _negoState.GetIdentity();
+        public virtual TokenImpersonationLevel ImpersonationLevel
+        {
+            get
+            {
+                ThrowIfFailed(authSuccessCheck: true);
+                return PrivateImpersonationLevel;
+            }
+        }
+
+        private TokenImpersonationLevel PrivateImpersonationLevel =>
+            _context!.IsDelegationFlag && _context.ProtocolName != NegotiationInfoClass.NTLM ? TokenImpersonationLevel.Delegation : // We should suppress the delegate flag in NTLM case.
+            _context.IsIdentifyFlag ? TokenImpersonationLevel.Identification :
+            TokenImpersonationLevel.Impersonation;
+
+        private bool HandshakeComplete => _context!.IsCompleted && _context.IsValidContext;
+
+        private bool CanGetSecureStream => _context!.IsConfidentialityFlag || _context.IsIntegrityFlag;
+
+        public virtual IIdentity RemoteIdentity
+        {
+            get
+            {
+                IIdentity? identity = _remoteIdentity;
+                if (identity is null)
+                {
+                    ThrowIfFailed(authSuccessCheck: true);
+                    _remoteIdentity = identity = NegotiateStreamPal.GetIdentity(_context!);
+                }
+                return identity;
+            }
+        }
 
         public override bool CanSeek => false;
 
@@ -271,8 +335,8 @@ namespace System.Net.Security
         {
             ValidateParameters(buffer, offset, count);
 
-            _negoState.CheckThrow(true);
-            if (!_negoState.CanGetSecureStream)
+            ThrowIfFailed(authSuccessCheck: true);
+            if (!CanGetSecureStream)
             {
                 return InnerStream.Read(buffer, offset, count);
             }
@@ -286,8 +350,8 @@ namespace System.Net.Security
         {
             ValidateParameters(buffer, offset, count);
 
-            _negoState.CheckThrow(true);
-            if (!_negoState.CanGetSecureStream)
+            ThrowIfFailed(authSuccessCheck: true);
+            if (!CanGetSecureStream)
             {
                 return InnerStream.ReadAsync(buffer, offset, count, cancellationToken);
             }
@@ -297,8 +361,8 @@ namespace System.Net.Security
 
         public override ValueTask<int> ReadAsync(Memory<byte> buffer, CancellationToken cancellationToken = default)
         {
-            _negoState.CheckThrow(true);
-            if (!_negoState.CanGetSecureStream)
+            ThrowIfFailed(authSuccessCheck: true);
+            if (!CanGetSecureStream)
             {
                 return InnerStream.ReadAsync(buffer, cancellationToken);
             }
@@ -340,7 +404,7 @@ namespace System.Net.Security
 
                     // The body carries 4 bytes for trailer size slot plus trailer, hence <= 4 frame size is always an error.
                     // Additionally we'd like to restrict the read frame size to 64k.
-                    if (readBytes <= 4 || readBytes > NegoState.MaxReadFrameSize)
+                    if (readBytes <= 4 || readBytes > MaxReadFrameSize)
                     {
                         throw new IOException(SR.net_frame_read_size);
                     }
@@ -362,7 +426,7 @@ namespace System.Net.Security
 
                     // Decrypt into internal buffer, change "readBytes" to count now _Decrypted Bytes_
                     // Decrypted data start from zero offset, the size can be shrunk after decryption.
-                    _bufferCount = readBytes = _negoState.DecryptData(_buffer!, 0, readBytes, out _bufferOffset);
+                    _bufferCount = readBytes = DecryptData(_buffer!, 0, readBytes, out _bufferOffset);
                     if (readBytes == 0 && buffer.Length != 0)
                     {
                         // Read again.
@@ -395,8 +459,8 @@ namespace System.Net.Security
         {
             ValidateParameters(buffer, offset, count);
 
-            _negoState.CheckThrow(true);
-            if (!_negoState.CanGetSecureStream)
+            ThrowIfFailed(authSuccessCheck: true);
+            if (!CanGetSecureStream)
             {
                 InnerStream.Write(buffer, offset, count);
                 return;
@@ -409,8 +473,8 @@ namespace System.Net.Security
         {
             ValidateParameters(buffer, offset, count);
 
-            _negoState.CheckThrow(true);
-            if (!_negoState.CanGetSecureStream)
+            ThrowIfFailed(authSuccessCheck: true);
+            if (!CanGetSecureStream)
             {
                 return InnerStream.WriteAsync(buffer, offset, count, cancellationToken);
             }
@@ -420,8 +484,8 @@ namespace System.Net.Security
 
         public override ValueTask WriteAsync(ReadOnlyMemory<byte> buffer, CancellationToken cancellationToken = default)
         {
-            _negoState.CheckThrow(true);
-            if (!_negoState.CanGetSecureStream)
+            ThrowIfFailed(authSuccessCheck: true);
+            if (!CanGetSecureStream)
             {
                 return InnerStream.WriteAsync(buffer, cancellationToken);
             }
@@ -441,11 +505,11 @@ namespace System.Net.Security
                 byte[]? outBuffer = null;
                 while (!buffer.IsEmpty)
                 {
-                    int chunkBytes = Math.Min(buffer.Length, NegoState.MaxWriteDataSize);
+                    int chunkBytes = Math.Min(buffer.Length, MaxWriteDataSize);
                     int encryptedBytes;
                     try
                     {
-                        encryptedBytes = _negoState.EncryptData(buffer.Slice(0, chunkBytes).Span, ref outBuffer);
+                        encryptedBytes = EncryptData(buffer.Slice(0, chunkBytes).Span, ref outBuffer);
                     }
                     catch (Exception e)
                     {
@@ -571,5 +635,611 @@ namespace System.Net.Security
                 throw new ArgumentOutOfRangeException(nameof(count), SR.net_offset_plus_count);
             }
         }
+
+        private void ValidateCreateContext(
+            string package,
+            NetworkCredential credential,
+            string servicePrincipalName,
+            ExtendedProtectionPolicy? policy,
+            ProtectionLevel protectionLevel,
+            TokenImpersonationLevel impersonationLevel)
+        {
+            if (policy != null)
+            {
+                // One of these must be set if EP is turned on
+                if (policy.CustomChannelBinding == null && policy.CustomServiceNames == null)
+                {
+                    throw new ArgumentException(SR.net_auth_must_specify_extended_protection_scheme, nameof(policy));
+                }
+
+                _extendedProtectionPolicy = policy;
+            }
+            else
+            {
+                _extendedProtectionPolicy = new ExtendedProtectionPolicy(PolicyEnforcement.Never);
+            }
+
+            ValidateCreateContext(package, true, credential, servicePrincipalName, _extendedProtectionPolicy!.CustomChannelBinding, protectionLevel, impersonationLevel);
+        }
+
+        private void ValidateCreateContext(
+            string package,
+            bool isServer,
+            NetworkCredential credential,
+            string? servicePrincipalName,
+            ChannelBinding? channelBinding,
+            ProtectionLevel protectionLevel,
+            TokenImpersonationLevel impersonationLevel)
+        {
+            if (_exception != null && !_canRetryAuthentication)
+            {
+                ExceptionDispatchInfo.Throw(_exception);
+            }
+
+            if (_context != null && _context.IsValidContext)
+            {
+                throw new InvalidOperationException(SR.net_auth_reauth);
+            }
+
+            if (credential == null)
+            {
+                throw new ArgumentNullException(nameof(credential));
+            }
+
+            if (servicePrincipalName == null)
+            {
+                throw new ArgumentNullException(nameof(servicePrincipalName));
+            }
+
+            NegotiateStreamPal.ValidateImpersonationLevel(impersonationLevel);
+            if (_context != null && IsServer != isServer)
+            {
+                throw new InvalidOperationException(SR.net_auth_client_server);
+            }
+
+            _exception = null;
+            _remoteOk = false;
+            _framer = new StreamFramer(InnerStream);
+            _framer.WriteHeader.MessageId = FrameHeader.HandshakeId;
+
+            _expectedProtectionLevel = protectionLevel;
+            _expectedImpersonationLevel = isServer ? impersonationLevel : TokenImpersonationLevel.None;
+            _writeSequenceNumber = 0;
+            _readSequenceNumber = 0;
+
+            ContextFlagsPal flags = ContextFlagsPal.Connection;
+
+            // A workaround for the client when talking to Win9x on the server side.
+            if (protectionLevel == ProtectionLevel.None && !isServer)
+            {
+                package = NegotiationInfoClass.NTLM;
+            }
+            else if (protectionLevel == ProtectionLevel.EncryptAndSign)
+            {
+                flags |= ContextFlagsPal.Confidentiality;
+            }
+            else if (protectionLevel == ProtectionLevel.Sign)
+            {
+                // Assuming user expects NT4 SP4 and above.
+                flags |= ContextFlagsPal.ReplayDetect | ContextFlagsPal.SequenceDetect | ContextFlagsPal.InitIntegrity;
+            }
+
+            if (isServer)
+            {
+                if (_extendedProtectionPolicy!.PolicyEnforcement == PolicyEnforcement.WhenSupported)
+                {
+                    flags |= ContextFlagsPal.AllowMissingBindings;
+                }
+
+                if (_extendedProtectionPolicy.PolicyEnforcement != PolicyEnforcement.Never &&
+                    _extendedProtectionPolicy.ProtectionScenario == ProtectionScenario.TrustedProxy)
+                {
+                    flags |= ContextFlagsPal.ProxyBindings;
+                }
+            }
+            else
+            {
+                // Server side should not request any of these flags.
+                if (protectionLevel != ProtectionLevel.None)
+                {
+                    flags |= ContextFlagsPal.MutualAuth;
+                }
+
+                if (impersonationLevel == TokenImpersonationLevel.Identification)
+                {
+                    flags |= ContextFlagsPal.InitIdentify;
+                }
+
+                if (impersonationLevel == TokenImpersonationLevel.Delegation)
+                {
+                    flags |= ContextFlagsPal.Delegate;
+                }
+            }
+
+            _canRetryAuthentication = false;
+
+            try
+            {
+                _context = new NTAuthentication(isServer, package, credential, servicePrincipalName, flags, channelBinding!);
+            }
+            catch (Win32Exception e)
+            {
+                throw new AuthenticationException(SR.net_auth_SSPI, e);
+            }
+        }
+
+        private Exception SetFailed(Exception e)
+        {
+            if (_exception == null || !(_exception is ObjectDisposedException))
+            {
+                _exception = e;
+            }
+
+            _context?.CloseContext();
+
+            return _exception;
+        }
+
+        private void ThrowIfFailed(bool authSuccessCheck)
+        {
+            if (_exception != null)
+            {
+                ExceptionDispatchInfo.Throw(_exception);
+            }
+
+            if (authSuccessCheck && !IsAuthenticatedCore)
+            {
+                throw new InvalidOperationException(SR.net_auth_noauth);
+            }
+        }
+
+        private void ProcessAuthentication(LazyAsyncResult? lazyResult)
+        {
+            ThrowIfFailed(authSuccessCheck: false);
+            if (Interlocked.Exchange(ref _authInProgress, 1) == 1)
+            {
+                throw new InvalidOperationException(SR.Format(SR.net_io_invalidnestedcall, lazyResult == null ? "BeginAuthenticate" : "Authenticate", "authenticate"));
+            }
+
+            try
+            {
+                if (_context!.IsServer)
+                {
+                    // Listen for a client blob.
+                    StartReceiveBlob(lazyResult);
+                }
+                else
+                {
+                    // Start with the first blob.
+                    StartSendBlob(null, lazyResult);
+                }
+            }
+            catch (Exception e)
+            {
+                e = SetFailed(e);
+                throw;
+            }
+            finally
+            {
+                if (lazyResult == null || _exception != null)
+                {
+                    _authInProgress = 0;
+                }
+            }
+        }
+
+        private void EndProcessAuthentication(IAsyncResult result)
+        {
+            if (result == null)
+            {
+                throw new ArgumentNullException("asyncResult");
+            }
+
+            LazyAsyncResult? lazyResult = result as LazyAsyncResult;
+            if (lazyResult == null)
+            {
+                throw new ArgumentException(SR.Format(SR.net_io_async_result, result.GetType().FullName), "asyncResult");
+            }
+
+            if (Interlocked.Exchange(ref _authInProgress, 0) == 0)
+            {
+                throw new InvalidOperationException(SR.Format(SR.net_io_invalidendcall, "EndAuthenticate"));
+            }
+
+            // No "artificial" timeouts implemented so far, InnerStream controls that.
+            lazyResult.InternalWaitForCompletion();
+
+            if (lazyResult.Result is Exception e)
+            {
+                e = SetFailed(e);
+                ExceptionDispatchInfo.Throw(e);
+            }
+        }
+
+        private bool CheckSpn()
+        {
+            if (_context!.IsKerberos ||
+                _extendedProtectionPolicy!.PolicyEnforcement == PolicyEnforcement.Never ||
+                _extendedProtectionPolicy.CustomServiceNames == null)
+            {
+                return true;
+            }
+
+            string? clientSpn = _context.ClientSpecifiedSpn;
+
+            if (string.IsNullOrEmpty(clientSpn))
+            {
+                return _extendedProtectionPolicy.PolicyEnforcement == PolicyEnforcement.WhenSupported;
+            }
+
+            return _extendedProtectionPolicy.CustomServiceNames.Contains(clientSpn);
+        }
+
+        //
+        // Client side starts here, but server also loops through this method.
+        //
+        private void StartSendBlob(byte[]? message, LazyAsyncResult? lazyResult)
+        {
+            Exception? exception = null;
+            if (message != s_emptyMessage)
+            {
+                message = GetOutgoingBlob(message, ref exception);
+            }
+
+            if (exception != null)
+            {
+                // Signal remote side on a failed attempt.
+                StartSendAuthResetSignal(lazyResult, message!, exception);
+                return;
+            }
+
+            if (HandshakeComplete)
+            {
+                if (_context!.IsServer && !CheckSpn())
+                {
+                    exception = new AuthenticationException(SR.net_auth_bad_client_creds_or_target_mismatch);
+                    int statusCode = ERROR_TRUST_FAILURE;
+                    message = new byte[sizeof(long)];
+
+                    for (int i = message.Length - 1; i >= 0; --i)
+                    {
+                        message[i] = (byte)(statusCode & 0xFF);
+                        statusCode = (int)((uint)statusCode >> 8);
+                    }
+
+                    StartSendAuthResetSignal(lazyResult, message, exception);
+                    return;
+                }
+
+                if (PrivateImpersonationLevel < _expectedImpersonationLevel)
+                {
+                    exception = new AuthenticationException(SR.Format(SR.net_auth_context_expectation, _expectedImpersonationLevel.ToString(), PrivateImpersonationLevel.ToString()));
+                    int statusCode = ERROR_TRUST_FAILURE;
+                    message = new byte[sizeof(long)];
+
+                    for (int i = message.Length - 1; i >= 0; --i)
+                    {
+                        message[i] = (byte)(statusCode & 0xFF);
+                        statusCode = (int)((uint)statusCode >> 8);
+                    }
+
+                    StartSendAuthResetSignal(lazyResult, message, exception);
+                    return;
+                }
+
+                ProtectionLevel result = _context.IsConfidentialityFlag ? ProtectionLevel.EncryptAndSign : _context.IsIntegrityFlag ? ProtectionLevel.Sign : ProtectionLevel.None;
+
+                if (result < _expectedProtectionLevel)
+                {
+                    exception = new AuthenticationException(SR.Format(SR.net_auth_context_expectation, result.ToString(), _expectedProtectionLevel.ToString()));
+                    int statusCode = ERROR_TRUST_FAILURE;
+                    message = new byte[sizeof(long)];
+
+                    for (int i = message.Length - 1; i >= 0; --i)
+                    {
+                        message[i] = (byte)(statusCode & 0xFF);
+                        statusCode = (int)((uint)statusCode >> 8);
+                    }
+
+                    StartSendAuthResetSignal(lazyResult, message, exception);
+                    return;
+                }
+
+                // Signal remote party that we are done
+                _framer!.WriteHeader.MessageId = FrameHeader.HandshakeDoneId;
+                if (_context.IsServer)
+                {
+                    // Server may complete now because client SSPI would not complain at this point.
+                    _remoteOk = true;
+
+                    // However the client will wait for server to send this ACK
+                    // Force signaling server OK to the client
+                    message ??= s_emptyMessage;
+                }
+            }
+            else if (message == null || message == s_emptyMessage)
+            {
+                throw new InternalException();
+            }
+
+            if (message != null)
+            {
+                //even if we are completed, there could be a blob for sending.
+                if (lazyResult == null)
+                {
+                    _framer!.WriteMessage(message);
+                }
+                else
+                {
+                    IAsyncResult ar = _framer!.BeginWriteMessage(message, s_writeCallback, lazyResult);
+                    if (!ar.CompletedSynchronously)
+                    {
+                        return;
+                    }
+                    _framer.EndWriteMessage(ar);
+                }
+            }
+
+            CheckCompletionBeforeNextReceive(lazyResult);
+        }
+
+        //
+        // This will check and logically complete the auth handshake.
+        //
+        private void CheckCompletionBeforeNextReceive(LazyAsyncResult? lazyResult)
+        {
+            if (HandshakeComplete && _remoteOk)
+            {
+                // We are done with success.
+                lazyResult?.InvokeCallback();
+                return;
+            }
+
+            StartReceiveBlob(lazyResult);
+        }
+
+        //
+        // Server side starts here, but client also loops through this method.
+        //
+        private void StartReceiveBlob(LazyAsyncResult? lazyResult)
+        {
+            Debug.Assert(_framer != null);
+
+            byte[]? message;
+            if (lazyResult == null)
+            {
+                message = _framer.ReadMessage();
+            }
+            else
+            {
+                IAsyncResult ar = _framer.BeginReadMessage(s_readCallback, lazyResult);
+                if (!ar.CompletedSynchronously)
+                {
+                    return;
+                }
+
+                message = _framer.EndReadMessage(ar);
+            }
+
+            ProcessReceivedBlob(message, lazyResult);
+        }
+
+        private void ProcessReceivedBlob(byte[]? message, LazyAsyncResult? lazyResult)
+        {
+            // This is an EOF otherwise we would get at least *empty* message but not a null one.
+            if (message == null)
+            {
+                throw new AuthenticationException(SR.net_auth_eof, null);
+            }
+
+            // Process Header information.
+            if (_framer!.ReadHeader.MessageId == FrameHeader.HandshakeErrId)
+            {
+                if (message.Length >= sizeof(long))
+                {
+                    // Try to recover remote win32 Exception.
+                    long error = 0;
+                    for (int i = 0; i < 8; ++i)
+                    {
+                        error = (error << 8) + message[i];
+                    }
+
+                    ThrowCredentialException(error);
+                }
+
+                throw new AuthenticationException(SR.net_auth_alert, null);
+            }
+
+            if (_framer.ReadHeader.MessageId == FrameHeader.HandshakeDoneId)
+            {
+                _remoteOk = true;
+            }
+            else if (_framer.ReadHeader.MessageId != FrameHeader.HandshakeId)
+            {
+                throw new AuthenticationException(SR.Format(SR.net_io_header_id, "MessageId", _framer.ReadHeader.MessageId, FrameHeader.HandshakeId), null);
+            }
+
+            CheckCompletionBeforeNextSend(message, lazyResult);
+        }
+
+        //
+        // This will check and logically complete the auth handshake.
+        //
+        private void CheckCompletionBeforeNextSend(byte[] message, LazyAsyncResult? lazyResult)
+        {
+            //If we are done don't go into send.
+            if (HandshakeComplete)
+            {
+                if (!_remoteOk)
+                {
+                    throw new AuthenticationException(SR.Format(SR.net_io_header_id, "MessageId", _framer!.ReadHeader.MessageId, FrameHeader.HandshakeDoneId), null);
+                }
+
+                lazyResult?.InvokeCallback();
+                return;
+            }
+
+            // Not yet done, get a new blob and send it if any.
+            StartSendBlob(message, lazyResult);
+        }
+
+        //
+        //  This is to reset auth state on the remote side.
+        //  If this write succeeds we will allow auth retrying.
+        //
+        private void StartSendAuthResetSignal(LazyAsyncResult? lazyResult, byte[] message, Exception exception)
+        {
+            _framer!.WriteHeader.MessageId = FrameHeader.HandshakeErrId;
+
+            if (IsLogonDeniedException(exception))
+            {
+                exception = new InvalidCredentialException(IsServer ? SR.net_auth_bad_client_creds : SR.net_auth_bad_client_creds_or_target_mismatch, exception);
+            }
+
+            if (!(exception is AuthenticationException))
+            {
+                exception = new AuthenticationException(SR.net_auth_SSPI, exception);
+            }
+
+            if (lazyResult == null)
+            {
+                _framer.WriteMessage(message);
+            }
+            else
+            {
+                lazyResult.Result = exception;
+                IAsyncResult ar = _framer.BeginWriteMessage(message, s_writeCallback, lazyResult);
+                if (!ar.CompletedSynchronously)
+                {
+                    return;
+                }
+
+                _framer.EndWriteMessage(ar);
+            }
+
+            _canRetryAuthentication = true;
+            ExceptionDispatchInfo.Throw(exception);
+        }
+
+        private static void WriteCallback(IAsyncResult transportResult)
+        {
+            if (!(transportResult.AsyncState is LazyAsyncResult))
+            {
+                NetEventSource.Fail(transportResult, "State type is wrong, expected LazyAsyncResult.");
+            }
+
+            if (transportResult.CompletedSynchronously)
+            {
+                return;
+            }
+
+            // Async completion.
+            LazyAsyncResult lazyResult = (LazyAsyncResult)transportResult.AsyncState!;
+            try
+            {
+                NegotiateStream authState = (NegotiateStream)lazyResult.AsyncObject!;
+                authState._framer!.EndWriteMessage(transportResult);
+
+                // Special case for an error notification.
+                if (lazyResult.Result is Exception e)
+                {
+                    authState._canRetryAuthentication = true;
+                    ExceptionDispatchInfo.Throw(e);
+                }
+
+                authState.CheckCompletionBeforeNextReceive(lazyResult);
+            }
+            catch (Exception e) when (!lazyResult.InternalPeekCompleted) // this will throw on a worker thread.
+            {
+                lazyResult.InvokeCallback(e);
+            }
+        }
+
+        private static void ReadCallback(IAsyncResult transportResult)
+        {
+            if (!(transportResult.AsyncState is LazyAsyncResult))
+            {
+                NetEventSource.Fail(transportResult, "State type is wrong, expected LazyAsyncResult.");
+            }
+
+            if (transportResult.CompletedSynchronously)
+            {
+                return;
+            }
+
+            // Async completion.
+            LazyAsyncResult lazyResult = (LazyAsyncResult)transportResult.AsyncState!;
+            try
+            {
+                NegotiateStream authState = (NegotiateStream)lazyResult.AsyncObject!;
+                byte[]? message = authState._framer!.EndReadMessage(transportResult);
+                authState.ProcessReceivedBlob(message, lazyResult);
+            }
+            catch (Exception e) when (!lazyResult.InternalPeekCompleted) // this will throw on a worker thread.
+            {
+                lazyResult.InvokeCallback(e);
+            }
+        }
+
+        private static bool IsError(SecurityStatusPal status) =>
+            (int)status.ErrorCode >= (int)SecurityStatusPalErrorCode.OutOfMemory;
+
+        private unsafe byte[]? GetOutgoingBlob(byte[]? incomingBlob, ref Exception? e)
+        {
+            byte[]? message = _context!.GetOutgoingBlob(incomingBlob, false, out SecurityStatusPal statusCode);
+
+            if (IsError(statusCode))
+            {
+                e = NegotiateStreamPal.CreateExceptionFromError(statusCode);
+                uint error = (uint)e.HResult;
+
+                message = new byte[sizeof(long)];
+                for (int i = message.Length - 1; i >= 0; --i)
+                {
+                    message[i] = (byte)(error & 0xFF);
+                    error >>= 8;
+                }
+            }
+
+            if (message != null && message.Length == 0)
+            {
+                message = s_emptyMessage;
+            }
+
+            return message;
+        }
+
+        private int EncryptData(ReadOnlySpan<byte> buffer, [NotNull] ref byte[]? outBuffer)
+        {
+            ThrowIfFailed(authSuccessCheck: true);
+
+            // SSPI seems to ignore this sequence number.
+            ++_writeSequenceNumber;
+            return _context!.Encrypt(buffer, ref outBuffer, _writeSequenceNumber);
+        }
+
+        private int DecryptData(byte[] buffer, int offset, int count, out int newOffset)
+        {
+            ThrowIfFailed(authSuccessCheck: true);
+
+            // SSPI seems to ignore this sequence number.
+            ++_readSequenceNumber;
+            return _context!.Decrypt(buffer, offset, count, out newOffset, _readSequenceNumber);
+        }
+
+        private static void ThrowCredentialException(long error)
+        {
+            var e = new Win32Exception((int)error);
+            throw e.NativeErrorCode switch
+            {
+                (int)SecurityStatusPalErrorCode.LogonDenied => new InvalidCredentialException(SR.net_auth_bad_client_creds, e),
+                ERROR_TRUST_FAILURE => new AuthenticationException(SR.net_auth_context_expectation_remote, e),
+                _ => new AuthenticationException(SR.net_auth_alert, e)
+            };
+        }
+
+        private static bool IsLogonDeniedException(Exception exception) =>
+            exception is Win32Exception win32exception &&
+            win32exception.NativeErrorCode == (int)SecurityStatusPalErrorCode.LogonDenied;
     }
 }
